Enum hyperloglog::Precision

#[repr(u8)]
pub enum Precision {

    P4,
    P5,
    P6,
    P7,
    P8,
    P9,
    P10,
    P11,
    P12,
    P13,
    P14,
    P15,
    P16,
    P17,
    P18,

}

The precision of the hyperloglog, that is, the number of bytes used for sharding accross registers.

A hyperloglog with precision p will have 2^p registers and require a proportional amount of memory.

Default precision is 12 bits, i.e. 4096 registers.

Examples

use hyperloglog::{HyperLogLog, Precision};

let hll = HyperLogLog::<i32>::with_precision(Precision::P12);

Variants

P4

4-bit precision.

P5

5-bit precision.

P6

6-bit precision.

P7

7-bit precision.

P8

8-bit precision.

P9

9-bit precision.

P10

10-bit precision.

P11

11-bit precision.

P12

12-bit precision (default).

P13

13-bit precision.

P14

14-bit precision.

P15

15-bit precision.

P16

16-bit precision.

P17

17-bit precision.

P18

18-bit precision.

Implementations

impl Precision

pub const MIN: Self = Self::P4

The smallest available precision: 4 bits.

Examples

use hyperloglog::Precision;

assert_eq!(Precision::MIN, Precision::P4);

pub const MAX: Self = Self::P18

The largest available precision: 18 bits.

Examples

use hyperloglog::Precision;

assert_eq!(Precision::MAX, Precision::P18);

pub const fn variants() -> &'static [Self]

All available precisions, in increasing order.

Examples

use hyperloglog::Precision;

assert!(Precision::variants().contains(&Precision::P12));

pub const fn in_range(value: u8) -> bool

Checks whether the given value is in the range of valid precisions.

Examples

use hyperloglog::Precision;

assert!(Precision::in_range(12));
assert!(!Precision::in_range(20));

pub const fn new(value: u8) -> Option<Self>

Creates a precision if the given value is in range.

Examples

use hyperloglog::Precision;

assert_eq!(Precision::new(12), Some(Precision::P12));
assert_eq!(Precision::new(20), None);

pub const fn get(self) -> u8

Return the precision value as an u8.

Examples

use hyperloglog::Precision;

assert_eq!(Precision::P12.get(), 12);