Struct hyperloglog::HyperLogLog

pub struct HyperLogLog<T, S = BuildHasherDefault<DefaultHasher>> where
    T: ?Sized,  { /* fields omitted */ }

A hyperloglog data structure to estimate the number of distinct elements in a data stream.

use hyperloglog::HyperLogLog;

let mut hll = HyperLogLog::<i32>::new();
for i in 0..100_000 {
     hll.insert(&i);
}
assert!((50_000..150_000).contains(&hll.len()));

Implementations

impl<T, S> HyperLogLog<T, S> where
T: ?Sized,
S: Default,

pub fn new() -> Self

Creates a new empty hyperloglog with the default precision.

Examples

use hyperloglog::HyperLogLog;

let hll = HyperLogLog::<i32>::new();

pub fn with_precision(precision: Precision) -> Self

Creates a new empty hyperloglog with the given precision.

Examples

use hyperloglog::{Precision, HyperLogLog};

let hll = HyperLogLog::<i32>::with_precision(Precision::P12);

impl<T, S> HyperLogLog<T, S> where
T: ?Sized,

pub fn with_hasher(hasher: S) -> Self

Creates a new empty hyperloglog with the default precision and the given hasher to hash keys.

Examples

use std::collections::hash_map::RandomState;

use hyperloglog::HyperLogLog;

let s = RandomState::new();
let mut hll = HyperLogLog::<i32, _>::with_hasher(s);
hll.insert(&1);

pub fn with_precision_and_hasher(precision: Precision, hasher: S) -> Self

Creates a new empty hyperloglog with the given precision and the given hasher to hash keys.

Examples

use std::collections::hash_map::RandomState;

use hyperloglog::{Precision, HyperLogLog};

let s = RandomState::new();
let mut hll = HyperLogLog::<i32, _>::with_precision_and_hasher(Precision::P12, s);
hll.insert(&1);

pub fn precision(&self) -> Precision

Returns the precision of the hyperloglog.

Examples

use hyperloglog::{HyperLogLog, Precision};

let mut hll = HyperLogLog::<i32>::with_precision(Precision::P12);
assert_eq!(hll.precision(), Precision::P12);

pub fn hasher(&self) -> &S

Returns a reference to the hyperloglog’s BuildHasher.

Examples

use std::collections::hash_map::RandomState;

use hyperloglog::HyperLogLog;

let hasher = RandomState::new();
let hll = HyperLogLog::<i32, _>::with_hasher(hasher);
let hasher: &RandomState = hll.hasher();

pub fn insert_hash(&mut self, h: u64)

Adds a hash value to the hyperloglog.

This may be handy when the hash is previously computed, to avoid computing twice. Hash values need to be uniformly distributed over u64 for an accurate total count.

In all other cases, use HyperLogLog::insert instead.

use std::{
    collections::hash_map::RandomState,
    hash::{BuildHasher, Hash, Hasher},
};

use hyperloglog::HyperLogLog;

let mut hasher = RandomState::new().build_hasher();
1i32.hash(&mut hasher);
let h = hasher.finish();

let mut hll = HyperLogLog::<i32, ()>::with_hasher(());
hll.insert_hash(h);

pub fn len(&self) -> usize

Calculates the approximate number of different elements.

Examples

use hyperloglog::HyperLogLog;

let mut hll = HyperLogLog::<i32>::new();
for i in 0..100_000 {
    hll.insert(&i);
}
assert!((50_000..150_000).contains(&hll.len()));

pub fn clear(&mut self)

Clears the hyperloglog, removing all values.

Examples

use hyperloglog::HyperLogLog;

let mut hll = HyperLogLog::<i32>::new();
hll.insert(&1);
hll.clear();
assert!(hll.is_empty());

pub fn is_empty(&self) -> bool

Returns true if the hyperloglog contains no elements.

Examples

use hyperloglog::HyperLogLog;

let mut hll = HyperLogLog::<i32>::new();
assert!(hll.is_empty());
hll.insert(&1);
assert!(!hll.is_empty());

pub fn merge_from_unchecked(&mut self, rhs: &Self)

Merges the hyperloglog rhs into self without checking that precisions and hashers are the same on both terms.

A checked variant is available via the try_merge_from method.

impl<T, S> HyperLogLog<T, S> where
T: ?Sized,
S: Clone,

pub fn merge_unchecked(&self, rhs: &Self) -> Self

Merges two hyperloglogs without checking that precisions and hashers are the same on both terms.

A checked variant is available via the try_merge method.

impl<T, S> HyperLogLog<T, S> where
T: ?Sized,
S: Eq,

pub fn try_merge_from(&mut self, rhs: &Self) -> Result<(), TryMergeError>

Merges the hyperloglog rhs into self.

Examples

use hyperloglog::HyperLogLog;

let mut hll1 = HyperLogLog::<i32>::new();
for i in 0..75_000 {
    hll1.insert(&i);
}
let mut hll2 = HyperLogLog::<i32>::new();
for i in 25_000..100_000 {
    hll2.insert(&i);
}
hll1.try_merge_from(&hll2).unwrap();
assert!((50_000..150_000).contains(&hll1.len()));

impl<T, S> HyperLogLog<T, S> where
T: ?Sized,
S: Clone + Eq,

pub fn try_merge(&self, rhs: &Self) -> Result<Self, TryMergeError>

Merges two hyperloglogs.

Examples

use hyperloglog::HyperLogLog;

let mut hll1 = HyperLogLog::<i32>::new();
for i in 0..75_000 {
    hll1.insert(&i);
}
let mut hll2 = HyperLogLog::<i32>::new();
for i in 25_000..100_000 {
    hll2.insert(&i);
}
let hll3 = hll1.try_merge(&hll2).unwrap();
assert!((50_000..150_000).contains(&hll3.len()));

impl<T, S> HyperLogLog<T, S> where
T: ?Sized,
S: BuildHasher,

pub fn insert<Q>(&mut self, value: &Q) where
T: Borrow<Q>,
Q: Hash + ?Sized,

Adds a value to the hyperloglog.

Examples

use hyperloglog::HyperLogLog;

let mut hll = HyperLogLog::<i32>::new();
hll.insert(&1);